requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
discord.py==1.7.3
python-dotenv==0.19.2
requests==2.26.0
